+++ yazi-scheduler/src/ongoing.rs
@@ -1,15 +1,15 @@
-use hashbrown::HashMap;
+use hashbrown::{HashMap, hash_map::RawEntryMut};
 use ordered_float::OrderedFloat;
 use yazi_config::YAZI;
 use yazi_parser::app::TaskSummary;
-use yazi_shared::{Id, Ids};
+use yazi_shared::{CompletionToken, Id, Ids};
 
 use super::Task;
-use crate::TaskProg;
+use crate::{TaskIn, TaskProg};
 
 #[derive(Default)]
 pub struct Ongoing {
-	pub(super) inner: HashMap<Id, Task>,
+	inner: HashMap<Id, Task>,
 }
 
 impl Ongoing {
@@ -23,10 +23,38 @@ impl Ongoing {
 		self.inner.entry(id).insert(Task::new::<T>(id, name)).into_mut()
 	}
 
+	pub(super) fn cancel(&mut self, id: Id) -> Option<TaskIn> {
+		match self.inner.raw_entry_mut().from_key(&id) {
+			RawEntryMut::Occupied(mut oe) => {
+				let task = oe.get_mut();
+				task.done.complete(false);
+
+				if let Some(hook) = task.hook.take() {
+					return Some(hook);
+				}
+
+				oe.remove();
+			}
+			RawEntryMut::Vacant(_) => {}
+		}
+		None
+	}
+
+	pub(super) fn fulfill(&mut self, id: Id) -> Option<Task> {
+		let task = self.inner.remove(&id)?;
+		task.done.complete(true);
+		Some(task)
+	}
+
 	#[inline]
 	pub fn get_mut(&mut self, id: Id) -> Option<&mut Task> { self.inner.get_mut(&id) }
 
 	pub fn get_id(&self, idx: usize) -> Option<Id> { self.values().nth(idx).map(|t| t.id) }
+
+	#[inline]
+	pub fn get_token(&self, id: Id) -> Option<CompletionToken> {
+		self.inner.get(&id).map(|t| t.done.clone())
+	}
 
 	pub fn len(&self) -> usize {
 		if YAZI.tasks.suppress_preload {
@@ -40,7 +68,9 @@ impl Ongoing {
 	pub fn exists(&self, id: Id) -> bool { self.inner.contains_key(&id) }
 
 	#[inline]
-	pub fn intact(&self, id: Id) -> bool { self.inner.get(&id).is_some_and(|t| !t.canceled) }
+	pub fn intact(&self, id: Id) -> bool {
+		self.inner.get(&id).is_some_and(|t| t.done.completed() != Some(false))
+	}
 
 	pub fn values(&self) -> Box<dyn Iterator<Item = &Task> + '_> {
 		if YAZI.tasks.suppress_preload {