+++ yazi-scheduler/src/process/out.rs
@@ -4,7 +4,6 @@ use crate::{Task, TaskProg};
 pub(crate) enum ProcessOutBlock {
 	Succ,
 	Fail(String),
-	Clean,
 }
 
 impl From<std::io::Error> for ProcessOutBlock {
@@ -22,9 +21,6 @@ impl ProcessOutBlock {
 				prog.state = Some(false);
 				task.log(reason);
 			}
-			Self::Clean => {
-				prog.cleaned = true;
-			}
 		}
 	}
 }
@@ -34,7 +30,6 @@ impl ProcessOutBlock {
 pub(crate) enum ProcessOutOrphan {
 	Succ,
 	Fail(String),
-	Clean,
 }
 
 impl From<anyhow::Error> for ProcessOutOrphan {
@@ -52,9 +47,6 @@ impl ProcessOutOrphan {
 				prog.state = Some(false);
 				task.log(reason);
 			}
-			Self::Clean => {
-				prog.cleaned = true;
-			}
 		}
 	}
 }
@@ -65,7 +57,6 @@ pub(crate) enum ProcessOutBg {
 	Log(String),
 	Succ,
 	Fail(String),
-	Clean,
 }
 
 impl From<anyhow::Error> for ProcessOutBg {
@@ -85,9 +76,6 @@ impl ProcessOutBg {
 			Self::Fail(reason) => {
 				prog.state = Some(false);
 				task.log(reason);
-			}
-			Self::Clean => {
-				prog.cleaned = true;
 			}
 		}
 	}