+++ yazi-scheduler/src/process/process.rs
@@ -55,14 +55,13 @@ impl Process {
 		})
 		.await?;
 
+		let done = task.done;
 		let mut stdout = BufReader::new(child.stdout.take().unwrap()).lines();
 		let mut stderr = BufReader::new(child.stderr.take().unwrap()).lines();
-		let mut cancel = task.cancel;
 		loop {
 			select! {
-				_ = cancel.recv() => {
+				false = done.future() => {
 					child.start_kill().ok();
-					cancel.close();
 					break;
 				}
 				Ok(Some(line)) = stdout.next_line() => {